requirements.txt
ultralytics==8.0.178
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
